[rpm] spec generators can use regexp filters
