+++ rpm/generate-spec-opensuse-generic.py
@@ -1,0 +1,8 @@
+#!/usr/bin/python
+
+import SpecGenerator
+
+SpecGenerator.process(
+    out_file = "statefs-providers-opensuse-generic.spec"
+    , filters = ['.*mce'])
+
